+++ src/tests/MidiNoteTest.cpp
@@ -74,7 +74,7 @@ class MidiNoteTest : public CppUnit::TestCase {
 		 * instruments sequential numbers starting from 36,
 		 * preserving legacy behavior. */
 
-		auto pSong = H2Core::Song::load( H2TEST_FILE( "song/test_song_0.9.6.h2song" ) );
+		auto pSong = H2Core::Song::load( H2TEST_FILE( "song/legacy/test_song_0.9.6.h2song" ) );
 		CPPUNIT_ASSERT( pSong != nullptr );
 
 		auto instruments = pSong->getInstrumentList();
@@ -95,7 +95,7 @@ class MidiNoteTest : public CppUnit::TestCase {
 		 * MIDI notes. Check that loading that song does not
 		 * change that mapping */
 
-		auto pSong = H2Core::Song::load( H2TEST_FILE( "song/test_song_0.9.7.h2song" ) );
+		auto pSong = H2Core::Song::load( H2TEST_FILE( "song/legacy/test_song_0.9.7.h2song" ) );
 		CPPUNIT_ASSERT( pSong != nullptr );
 
 		auto instruments = pSong->getInstrumentList();